foxy-system/apps/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
blessed==1.20.0
certifi==2024.8.30
charset-normalizer==3.3.2
easyocr==1.7.1
editor==1.6.6
exceptiongroup==1.2.2
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.9.0
h11==0.14.0
idna==3.8
imageio==2.35.1
inquirer==3.4.0
Jinja2==3.1.4
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
ninja==1.11.1.1
numpy==2.1.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pika==1.3.2
pillow==10.3.0
PyAutoGUI==0.9.54
pyclipper==1.3.0.post5
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.1
PyYAML==6.0.2
readchar==4.2.0
requests==2.31.0
rich==13.8.0
runs==1.2.2
scikit-image==0.24.0
scipy==1.14.1
selenium==4.18.1
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.13.2
tifffile==2024.8.30
torch==2.4.1
torchvision==0.19.1
trio==0.26.2
trio-websocket==0.11.1
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
wsproto==1.2.0
xmod==1.8.1
xvfbwrapper==0.2.9
yarl==1.10.0
